requirements.txt
streamlit==1.31.0
openai==1.12.0
Pillow==10.1.0
requests==2.31.0